feat: palette storybook 추가
