+++ src/components/groupList/RandomImage.tsx
@@ -16,7 +16,7 @@ const RandomImage = () => {
     <>
       {RandomData?.length ? (
         <Carousel
-          className='w-full' //max-w-sm
+          className='w-full max-w-sm'
           plugins={[
             Autoplay({
               delay: 3000,
